+++ ParaViewCore/ServerManager/Core/vtkSMProxyListDomain.cxx
@@ -75,7 +75,7 @@ void vtkSMProxyListDomain::AddProxy(const char* group, const char* name)
 //-----------------------------------------------------------------------------
 unsigned int vtkSMProxyListDomain::GetNumberOfProxyTypes()
 {
-  return this->Internals->ProxyTypeList.size();
+  return static_cast<unsigned int>(this->Internals->ProxyTypeList.size());
 }
 
 //-----------------------------------------------------------------------------
@@ -161,7 +161,7 @@ void vtkSMProxyListDomain::AddProxy(vtkSMProxy* proxy)
 //-----------------------------------------------------------------------------
 unsigned int vtkSMProxyListDomain::GetNumberOfProxies()
 {
-  return this->Internals->ProxyList.size();
+  return static_cast<unsigned int>(this->Internals->ProxyList.size());
 }
 
 //-----------------------------------------------------------------------------
